Python 3.10 support.

+++ pyfr/quadrules/__init__.py
@@ -38,8 +38,8 @@ class BaseTabulatedQuadRule(object):
             pts = [p[0] for p in pts]
 
         # Cast
-        self.pts = np.array(pts, dtype=np.float)
-        self.wts = np.array(wts, dtype=np.float)
+        self.pts = np.array(pts)
+        self.wts = np.array(wts)
 
 
 class BaseStoredQuadRule(BaseTabulatedQuadRule):
